# mB.Api/Features/Products/Commands/UpdateProduct/UpdateProductCommandResponse.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Threading.Tasks;

namespace mB.Api.Features.Products.Commands.UpdateProduct
{
    public class UpdateProductCommandResponse
    {
    }
}