# mB.Consumer/OpenAPIs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "mB.Api",
    "version": "v1"
  },
  "paths": {
    "/invoices": {
      "get": {
        "tags": [
          "Invoices"
        ],
        "operationId": "GetInvoices",
        "parameters": [
          {
            "name": "clientId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GetInvoicesQueryResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetInvoicesQueryResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetInvoicesQueryResponse"
                }
              }
            }
          }
        }
      }
    },
    "/invoices/add": {
      "post": {
        "tags": [
          "Invoices"
        ],
        "operationId": "AddInvoice",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddInvoiceCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AddInvoiceCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AddInvoiceCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AddInvoiceCommandResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddInvoiceCommandResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddInvoiceCommandResponse"
                }
              }
            }
          }
        }
      }
    },
    "/products": {
      "get": {
        "tags": [
          "Products"
        ],
        "operationId": "GetAllProducts",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GetProductsQueryResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetProductsQueryResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetProductsQueryResponse"
                }
              }
            }
          }
        }
      }
    },
    "/products/load": {
      "post": {
        "tags": [
          "Products"
        ],
        "operationId": "LoadProducts",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoadProductsCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LoadProductsCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/LoadProductsCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/LoadProductsCommandResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoadProductsCommandResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoadProductsCommandResponse"
                }
              }
            }
          }
        }
      }
    },
    "/products/update": {
      "post": {
        "tags": [
          "Products"
        ],
        "operationId": "UpdateProduct",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateProductCommandResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateProductCommandResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateProductCommandResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ClientDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "PositionDto": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "integer",
            "format": "int32"
          },
          "productName": {
            "type": "string",
            "nullable": true
          },
          "count": {
            "type": "integer",
            "format": "int32"
          },
          "cost": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
      "InvoiceDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "added": {
            "type": "string",
            "format": "date-time"
          },
          "paid": {
            "type": "boolean"
          },
          "client": {
            "$ref": "#/components/schemas/ClientDto"
          },
          "positions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PositionDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "GetInvoicesQueryResponse": {
        "type": "object",
        "properties": {
          "invoices": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InvoiceDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "AddInvoicePositionDto": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "integer",
            "format": "int32"
          },
          "count": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "AddInvoiceCommand": {
        "type": "object",
        "properties": {
          "clientId": {
            "type": "integer",
            "format": "int32"
          },
          "positions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AddInvoicePositionDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "AddInvoiceCommandResponse": {
        "type": "object",
        "properties": {
          "invoiceId": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "ProductDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "added": {
            "type": "string",
            "format": "date-time"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "category": {
            "type": "string",
            "nullable": true
          },
          "price": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
      "GetProductsQueryResponse": {
        "type": "object",
        "properties": {
          "products": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "LoadSource": {
        "enum": [
          0,
          1
        ],
        "type": "integer",
        "format": "int32"
      },
      "LoadProductsCommand": {
        "type": "object",
        "properties": {
          "source": {
            "$ref": "#/components/schemas/LoadSource"
          }
        },
        "additionalProperties": false
      },
      "LoadProductsCommandResponse": {
        "type": "object",
        "properties": {
          "loadedProductIds": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "UpdateProductCommand": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "category": {
            "type": "string",
            "nullable": true
          },
          "price": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
      "UpdateProductCommandResponse": {
        "type": "object",
        "additionalProperties": false
      }
    }
  }
}